requirements.txt
Flask==3.0.0
mysql-connector-python==9.0.0
werkzeug==3.0.1
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.9
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
plotly==5.24.1

# spaCy is skipped due to Python 3.13 compatibility issues
# Uncomment the line below if using Python 3.12 or lower
# spacy==3.7.2
